Got basic routes working in URL
